+++ src/Wallet.php
@@ -10,6 +10,8 @@ class Wallet
     private $arguments;
     private $validatorAddress;
     private $mnemonics;
+    public $currentNonce;
+    public $currentNonceValidUntil;
 
     /**
      * Wallet constructor.
@@ -19,12 +21,18 @@ class Wallet
      */
     //"m/44'/60'/0'/0/0"
     //how it was "m/44'/60'/0'/0"
-    public function __construct($mnemonics = null, $path_key = "m/44'/60'/0'/0")
-    {
+    public function __construct(
+        $mnemonics = null,
+        $path_key = "m/44'/60'/0'/0",
+        $currentNonce = null,
+        $currentNonceValidUntil = null
+    ){
         if (!$mnemonics) $mnemonics = WalletHelpers::createNewMnemonics();
         $this->mnemonics = $mnemonics;
         $this->arguments = WalletHelpers::generateNewAddress('dx', $mnemonics, $path_key);
         $this->validatorAddress = WalletHelpers::generateNewAddress('dxvaloper', $mnemonics, $path_key);
+        $this->currentNonce = $currentNonce;
+        $this->currentNonceValidUntil = $currentNonceValidUntil;
     }
 
     public function getAddress()
@@ -63,5 +71,11 @@ class Wallet
     public function getMnemonics()
     {
         return $this->mnemonics;
+    }
+
+    public function updateNonce($nonce)
+    {
+        $this->currentNonce = $nonce;
+        $this->currentNonceValidUntil = $nonce ? time(): null;
     }
 }